# src/swagger/swagger.json
{
	"basePath": "/",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"ProductCreateRequest": {
			"properties": {
				"id": {
					"type": "string"
				},
				"name": {
					"type": "string"
				},
				"desc": {
					"type": "string"
				},
				"price": {
					"type": "number",
					"format": "double"
				},
				"imageUrl": {
					"type": "string"
				}
			},
			"required": [
				"id",
				"name",
				"desc",
				"price",
				"imageUrl"
			],
			"type": "object"
		}
	},
	"info": {
		"title": "catalog-service",
		"version": "1.0.0",
		"description": "Product Catalog APIs",
		"license": {
			"name": "ISC"
		}
	},
	"paths": {
		"/api/products": {
			"get": {
				"operationId": "GetAll",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "object"
						}
					}
				},
				"description": "Get the all product",
				"security": [],
				"parameters": [
					{
						"default": 0,
						"in": "query",
						"name": "currentPage",
						"required": false,
						"format": "double",
						"type": "number"
					},
					{
						"in": "query",
						"name": "highPrice",
						"required": false,
						"format": "double",
						"type": "number"
					}
				]
			},
			"post": {
				"operationId": "Create",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "object"
						}
					}
				},
				"description": "Create a product",
				"security": [],
				"parameters": [
					{
						"description": "This is a product creation request description",
						"in": "body",
						"name": "request",
						"required": true,
						"schema": {
							"$ref": "#/definitions/ProductCreateRequest"
						}
					}
				]
			}
		},
		"/api/products/{productId}": {
			"get": {
				"operationId": "Get",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "object"
						}
					}
				},
				"description": "Get product by Id",
				"security": [],
				"parameters": [
					{
						"description": "Product Id",
						"in": "path",
						"name": "productId",
						"required": true,
						"type": "string"
					}
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {}
}